+++ scripts/tasks/task-sysex-logging.py
@@ -2,7 +2,6 @@
 import sys
 import time
 import argparse
-import rtmidi
 import sys
 import util
 
@@ -101,6 +100,13 @@ def sysex_console(midiout, midiin):
 
 
 def main():
+    try:
+        import rtmidi
+    except ImportError:
+        util.install_rtmidi()
+    finally:
+        import rtmidi
+
     midiout = rtmidi.MidiOut()
     midiin = rtmidi.MidiIn()