+++ src/deluge/gui/menu_item/arpeggiator/preset_mode.h
@@ -67,7 +67,8 @@ public:
 		soundEditor.currentArpSettings->flagForceArpRestart = true;
 	}
 
-	deluge::vector<std::string_view> getOptions() override {
+	deluge::vector<std::string_view> getOptions(OptType optType) override {
+		(void)optType;
 		using enum l10n::String;
 		return {
 		    l10n::getView(STRING_FOR_OFF),    //<