+++ tests/32bit_unit_tests/memory_tests.cpp
@@ -207,8 +207,9 @@ TEST(MemoryAllocation, allocationSizes) {
 		for (int i = 0; i < expectedAllocations; i++) {
 			if (!testAllocations[i]) {
 				// this is to make a log distribution - probably the worst case for packing efficiency
-				int magnitude = rand() % 16;
-				int size = (rand() % 10) << magnitude;
+				// min allocation size is 8 bytes
+				int magnitude = rand() % 16 + 2;
+				int size = (rand() % 10 + 1) << magnitude;
 				void* testalloc = memreg.alloc(size, false, NULL);
 				if (testalloc) {
 					totalSize += size;
@@ -229,11 +230,13 @@ TEST(MemoryAllocation, allocationSizes) {
 
 		// we should have one empty space left, and it should be the size of the memory minus headers
 		CHECK(memreg.emptySpaces.getNumElements() == 1);
-		CHECK(memreg.emptySpaces.getKeyAtIndex(0) == mem_size - 16);
+		// we might have needed to align the region start to 16 after setting the headers
+		int sizeDiff = (mem_size - 16 - memreg.emptySpaces.getKeyAtIndex(0));
+		CHECK(sizeDiff <= 16);
 	}
 	// un modified GMA gets .999311
 	// current with extra padding gets .9939
-	// std::cout << "Packing factor: " << (average_packing_factor / numRepeats) << std::endl;
+	std::cout << "Packing factor: " << (average_packing_factor / numRepeats) << std::endl;
 	CHECK(average_packing_factor / numRepeats > 0.99);
 };